Add SingleAlbum component
