+++ client/components/ArtistView.js
@@ -2,7 +2,8 @@ import React, { Component } from 'react'
 import AllAlbums from "./AllAlbums"
 import PopularityMap from "./PopularityMap";
 import { connect } from 'react-redux'
-import { fetchOneArtist } from "../store/artist";
+import { fetchOneArtist } from "../store/artist"
+import axios from 'axios'
 
 
 import { withStyles } from '@material-ui/core/styles'
@@ -15,22 +16,41 @@ const styles = {
     },
 }
 
-
 class ArtistView extends Component {
     constructor(props) {
         super(props)
         this.state = {
-            //something
+            likes: [],
+            dislikes: [],
+            artist: {}
         }
+    }
+
+    async componentDidMount() {
+       const result = await axios.get(`/api/artists/${3}`)
+       this.setState({
+        artist: result.data
+    })
+
+        const { data } = await axios.get(`/api/likes/${3}`)
+        this.setState({
+            likes: data
+        })
+
+        const res = await axios.get(`/api/dislikes/${3}`)
+        this.setState({
+            dislikes: res.data
+        })
     }
 
     render() {
         const {classes} = this.props
+        const {likes, dislikes, artist} = this.state
 
         return (
             <div className={classes.container}>
-                <AllAlbums />
-                <PopularityMap />
+                <AllAlbums artist={artist} />
+                <PopularityMap likes={likes} dislikes={dislikes} />
             </div>
         )
     }
