+++ client/components/PopularityMap.js
@@ -28,22 +28,9 @@ class PopularityMap extends PureComponent {
     constructor() {
         super()
         this.state = {
-            likes: [],
-            dislikes: [],
             likesClicked: false,
             disLikesClicked: false
         }
-    }
-    async componentDidMount() {
-        const { data } = await axios.get(`/api/likes/${3}`)
-        this.setState({
-            likes: data
-        })
-
-        const res = await axios.get(`/api/dislikes/${3}`)
-        this.setState({
-            dislikes: res.data
-        })
     }
 
     onMarkerClick = (evt) => {
@@ -60,8 +47,8 @@ class PopularityMap extends PureComponent {
 
     render() {
         const getProvider = (x, y, z) => `https://cartodb-basemaps-a.global.ssl.fastly.net/light_all/${z}/${x}/${y}.png`;
-        const { classes } = this.props
-        const { likes, dislikes, likesClicked } = this.state
+        const { classes, likes, dislikes } = this.props
+        const { likesClicked } = this.state
         const mapConfig = {
             center: [41.850033, -87.6500523],
             zoom: 4
