+++ client/routes.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react'
 import { connect } from 'react-redux'
 import { withRouter, Route, Switch } from 'react-router-dom'
 import PropTypes from 'prop-types'
-import { Login, Signup, UserHome, Player, ArtistView, SingleAlbumView } from './components'
+import {Player, ArtistView, SingleAlbumView } from './components'
 import { me } from './store'
 import testIterator from './components/testIterator';
